+++ apps/nextjs/src/app/_components/posts.tsx
@@ -3,6 +3,7 @@
 import { use } from "react";
 
 import type { RouterOutputs } from "@acme/api";
+import { createPostSchema } from "@acme/db";
 import { cn } from "@acme/ui";
 import { Button } from "@acme/ui/button";
 import {
@@ -15,13 +16,12 @@ import {
 } from "@acme/ui/form";
 import { Input } from "@acme/ui/input";
 import { toast } from "@acme/ui/toast";
-import { CreatePostSchema } from "@acme/validators";
 
 import { api } from "~/trpc/react";
 
 export function CreatePostForm() {
   const form = useForm({
-    schema: CreatePostSchema,
+    schema: createPostSchema,
     defaultValues: {
       content: "",
       title: "",
