+++ packages/api/src/router/post.ts
@@ -1,8 +1,10 @@
 import type { TRPCRouterRecord } from "@trpc/server";
 import { z } from "zod";
 
-import { desc, eq, schema } from "@acme/db";
-import { CreatePostSchema } from "@acme/validators";
+import { createPostSchema, desc, eq } from "@acme/db";
+import { post } from "@acme/db/schema";
+
+// import { CreatePostSchema } from "@acme/validators";
 
 import { protectedProcedure, publicProcedure } from "../trpc";
 
@@ -10,7 +12,7 @@ export const postRouter = {
   all: publicProcedure.query(({ ctx }) => {
     // return ctx.db.select().from(schema.post).orderBy(desc(schema.post.id));
     return ctx.db.query.post.findMany({
-      orderBy: desc(schema.post.id),
+      orderBy: desc(post.id),
       limit: 10,
     });
   }),
@@ -24,17 +26,17 @@ export const postRouter = {
       //   .where(eq(schema.post.id, input.id));
 
       return ctx.db.query.post.findFirst({
-        where: eq(schema.post.id, input.id),
+        where: eq(post.id, input.id),
       });
     }),
 
   create: protectedProcedure
-    .input(CreatePostSchema)
+    .input(createPostSchema)
     .mutation(({ ctx, input }) => {
-      return ctx.db.insert(schema.post).values(input);
+      return ctx.db.insert(post).values(input);
     }),
 
   delete: protectedProcedure.input(z.number()).mutation(({ ctx, input }) => {
-    return ctx.db.delete(schema.post).where(eq(schema.post.id, input));
+    return ctx.db.delete(post).where(eq(post.id, input));
   }),
 } satisfies TRPCRouterRecord;
