+++ packages/db/drizzle.config.ts
@@ -1,0 +1,10 @@
+import type { Config } from "drizzle-kit";
+
+import { getConnectionString } from "./src/client";
+
+export default {
+  schema: "./src/schema",
+  driver: "mysql2",
+  dbCredentials: { uri: getConnectionString() },
+  tablesFilter: ["t3turbo_*"],
+} satisfies Config;
