--- packages/db/src/schema/_table.ts
@@ -1,9 +1,0 @@
-import { mysqlTableCreator } from "drizzle-orm/mysql-core";
-
-/**
- * This is an example of how to use the multi-project schema feature of Drizzle ORM.
- * Use the same database instance for multiple projects.
- *
- * @see https://orm.drizzle.team/docs/goodies#multi-project-schema
- */
-export const mySqlTable = mysqlTableCreator((name) => `t3turbo_${name}`);
